% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{af_dependency}
\alias{af_dependency}
\title{Attach dependency}
\usage{
af_dependency()
}
\description{
Attach aframe forcegraph dependency.
}
\note{
Only useful when using \code{aforce} in conjunction with other aframe packages.
}
