% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{data}
\name{aForce}
\alias{aForce}
\title{aforce R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
aForce
}
\description{
Create a VR force directed graph.
}
\section{Methods}{

\itemize{
  \item{\code{set_nodes} Add nodes.}
  \item{\code{set_edges} Add edges.}
  \item{\code{plot} Plot graph}
}
}

\keyword{datasets}
