% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{some_links}
\alias{some_links}
\title{Links}
\format{An igraph object with 181 nodes and 4 variables
\describe{
  \item{\code{source}}{ combination of letter and number.}
  \item{\code{target}}{ combination of letter and number.}
}}
\usage{
some_links
}
\description{
Randomly generated links for examples.
}
\keyword{datasets}
