% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{some_nodes}
\alias{some_nodes}
\title{Nodes}
\format{An igraph object with 181 nodes and 2 variables
\describe{
  \item{\code{id}}{ combination of letter and number.}
  \item{\code{label}}{ combination of letter and number.}
  \item{\code{val}}{ random value.}
  \item{\code{color}}{ random color.}
}}
\usage{
some_nodes
}
\description{
Randomly generated nodes for examples.
}
\keyword{datasets}
