% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{data}
\name{aForce}
\alias{aForce}
\title{aforce R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
aForce
}
\description{
Create a VR force directed graph.
}
\section{Methods}{

\itemize{
  \item{\code{nodes} Add nodes.}
  \item{\code{edges} Add edges.}
  \item{\code{build} Build graph}
  \item{\code{get} Get graph}
  \item{\code{browse} Browse graph}
  \item{\code{embed} Embed graph}
}
}

\examples{
if(interactive()){
  graph <- aForce$
    new()$
    nodes(some_nodes, id, val, color)$
    links(some_links, source, target)$
    build()$
    get()

  aframer::browse_aframe(graph)
}

}
\seealso{
\href{https://github.com/vasturiano/aframe-forcegraph-component}{official page}.
}
\keyword{datasets}
